requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
chromadb==0.5.0
openai==1.50.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
